Remove implicit cast to fix second conversion

+++ app/src/main/java/de/philipphager/disclosure/util/time/Stopwatch.java
@@ -32,7 +32,7 @@ public class Stopwatch {
   }
 
   private double toSecs() {
-    return duration.toMillis() / MILLIS;
+    return duration.toMillis() / (double) MILLIS;
   }
 
   @Override public String toString() {
